HUGE refactor with plenty fixes
