+++ package.json
@@ -17,12 +17,14 @@
     "@types/express": "^4.17.2",
     "@types/jsonwebtoken": "^8.3.5",
     "@types/mongoose": "^5.5.34",
+    "@types/uuid": "^3.4.6",
     "nodemon": "^2.0.1",
     "ts-node-dev": "^1.0.0-pre.44",
     "tsc": "^1.20150623.0",
     "typescript": "^3.7.3"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-async-handler": "^1.1.4",
@@ -30,6 +32,6 @@
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.3.5",
     "mongoose": "^5.7.13",
-    "body-parser": "^1.19.0"
+    "uuid": "^3.3.3"
   }
 }
