--- src/core/Service.ts
@@ -1,15 +1,0 @@
-import ServicePort from "./_ServicePort";
-import RepositoryPort from "./_ParticipantRepositoryPort";
-
-export default class Service implements ServicePort {
-    private _repository: RepositoryPort;
-
-    constructor(testRepository: RepositoryPort) {
-        this._repository = testRepository;
-    }
-
-    async findTestResult(participantId: String) {
-        const result = await this._repository.selectByTestId(participantId);
-        return result;
-    }
-}
