--- src/core/_GroupRepositoryPort.ts
@@ -1,8 +1,0 @@
-import GroupResult from "./domain/GroupResult";
-
-/**
- * An abstract class describing the expected behavior of TestRepository implementations
- */
-export default interface GroupRepositoryPort {
-    selectByTestId(id: String): Promise<GroupResult>;
-}
