--- src/core/domain/ParticipantResult.ts
@@ -1,7 +1,0 @@
-export default class ParticipantResult{
-    stringFromDb: String;
-
-    constructor(text: String){
-        this.stringFromDb = text;
-    }
-}
